Modify USBtoSerial for Adafruit Atmega32u4 Breakout
By default this modified makefile uses the built-in bootloader
installed on the breakout and defaults to /dev/ttyACM0 (Linux).

This is equivalent to:

make avrdude AVRDUDE_PROGRAMMER=avr109

If you want to use a USBtinyISP (connected to the breakout's ISP header):

make avrdude AVRDUDE_PROGRAMMER=usbtiny

To program (using the built-in bootloader):

1. Plug in your breakout over USB.
2. Press the white button.

$ make avrdude AVRDUDE_PORT=/dev/ttyACM0

It may help to type the line out ahead of time so that the bootloader
does not time out while waiting for an upload.

Using:

To get serial output, connect the device you want to get data from to
pin D2 (RXD1 on the Atmel datasheet) and power it on.

Data should come over the specified interface (for Linux this is usually
/dev/ttyACM0).